+++ domaindriven-server/src/DomainDriven.hs
@@ -1,4 +1,14 @@
-module DomainDriven
+module DomainDriven (module X) where
+
+import Data.UUID as X (UUID)
+
+import DomainDriven.Persistance.Class as X
+    ( ReadModel (..)
+    , Stored (..)
+    , WriteModel (..)
+    , mkId
+    )
+import DomainDriven.Server.Class as X
     ( Action
     , ActionHandler
     , ActionRunner
@@ -6,24 +16,20 @@ module DomainDriven
     , CbCmd
     , Cmd
     , HandlerType (..)
+    , ModelAccess (..)
+    , Query
+    , RequestType
     , mapEvent
     , mapModel
     , mapResult
-    , mkId
-    , ModelAccess (..)
-    , NamedJsonFields (..)
-    , Query
-    , ReadModel (..)
-    , RequestType
     , runAction
-    , Stored (..)
-    , UUID
-    , WriteModel (..)
     )
-where
-
-import Data.UUID
-import DomainDriven.Internal.Class
-import DomainDriven.Internal.NamedJsonFields
-    ( NamedJsonFields (..)
+import DomainDriven.Server.Config as X
+    ( ServerConfig
+    , mkServerConfig
+    )
+import DomainDriven.Server.TH as X (mkServer)
+import DomainDriven.Server.Types as X
+    ( ApiOptions (..)
+    , defaultApiOptions
     )
