--- domaindriven-server/src/DomainDriven/Server.hs
@@ -1,12 +1,0 @@
-{-# LANGUAGE DuplicateRecordFields #-}
-
-module DomainDriven.Server
-    ( mkServer
-    , module X
-    )
-where
-
-import DomainDriven.Server.Config as X
-import DomainDriven.Server.Param as X
-import DomainDriven.Server.TH
-import DomainDriven.Server.Types as X
