+++ domaindriven/domaindriven.cabal
@@ -27,11 +27,7 @@ source-repository head
 
 library
   exposed-modules:
-      DomainDriven.Internal.Class
-      DomainDriven.Internal.HasFieldName
-      DomainDriven.Internal.NamedFields
-      DomainDriven.Internal.NamedJsonFields
-      DomainDriven.Internal.Text
+      DomainDriven.Persistance.Class
       DomainDriven.Persistance.ForgetfulInMemory
       DomainDriven.Persistance.Postgres
       DomainDriven.Persistance.Postgres.Internal
@@ -86,13 +82,10 @@ library
     , http-types >=0.12.3
     , microlens >=0.4.12.0
     , mtl >=2.2.2
-    , openapi3 >=3.2.2
     , postgresql-simple >=0.6.4
     , random >=1.2.1.1
-    , servant-server >=0.19.2
     , streamly >=0.8.1.1
     , template-haskell >=2.18.0.0
-    , text >=1.2.5.0
     , time >=1.11.1.1
     , transformers >=0.5.6.2
     , unliftio >=0.2.0.1
@@ -145,20 +138,11 @@ test-suite domaindriven-test
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints -Wincomplete-record-updates -Wincomplete-patterns -threaded -rtsopts -with-rtsopts=-N -Wall -Wunused-packages
   build-depends:
       aeson >=2.0.3.0
-    , async >=2.2.4
     , base >=4.7 && <5
-    , containers >=0.6.5.1
-    , deepseq >=1.4.6.1
     , domaindriven
-    , exceptions >=0.10.4
-    , generic-lens >=2.2.1.0
     , hspec >=2.9.7
-    , microlens >=0.4.12.0
-    , openapi3 >=3.2.2
     , postgresql-simple >=0.6.4
-    , servant-server >=0.19.2
     , streamly >=0.8.1.1
-    , text >=1.2.5.0
     , time >=1.11.1.1
     , unliftio-pool >=0.2.2.0
     , uuid >=1.3.15
