+++ domaindriven/package.yaml
@@ -22,12 +22,6 @@ description:         Please see the README on GitHub at <https://github.com/tomm
 dependencies:
 - aeson >= 2.0.3.0
 - base >= 4.7 && < 5
-- containers >= 0.6.5.1
-- deepseq >= 1.4.6.1
-- exceptions >= 0.10.4
-- openapi3 >= 3.2.2
-- servant-server >= 0.19.2
-- text >= 1.2.5.0
 
 
 default-extensions:
@@ -85,6 +79,9 @@ library:
     - unordered-containers >= 0.2.19.1
     - vector >= 0.12.3.1
     - unliftio >= 0.2.0.1
+    - containers >= 0.6.5.1
+    - deepseq >= 1.4.6.1
+    - exceptions >= 0.10.4
 
     - streamly >= 0.8.1.1
     - postgresql-simple >= 0.6.4
@@ -107,11 +104,8 @@ tests:
     - -Wall
     - -Wunused-packages
     dependencies:
-    - async >= 2.2.4
     - domaindriven
-    - generic-lens >= 2.2.1.0
     - hspec >= 2.9.7
-    - microlens >= 0.4.12.0
     - postgresql-simple >= 0.6.4
     - streamly >= 0.8.1.1
     - time >=  1.11.1.1
